Menu torneio pronto a apresentar

+++ Classes/Utils.hpp
@@ -319,8 +319,8 @@ string normalizeId(unsigned int digits, unsigned int id);
 #endif /* Utils_hpp */
 
 template <class T>
-size_t findInVector(vector<T> vec1, T element) {
-	for (size_t i = 0; i < vec1.size(); i++) {
+int findInVector(vector<T> vec1, T element) {
+	for (int i = 0; i < vec1.size(); i++) {
 		if (element == vec1.at(i))
 			return i;
 	}
